+++ src/AppForSEII2526.API/DTOs/PurchaseDTO/PurchaseForDetailDTO.cs
@@ -4,7 +4,7 @@ namespace AppForSEII2526.API.DTOs.PurchaseDTO
 {
     public class PurchaseForDetailDTO : PurchaseForCreateDTO
     {
-        public PurchaseForDetailDTO(int id, string applicationUserName, string applicationUserSurname, string deliveryAddress, DateTime purchaseDate, double totalPrice, IList<PurchaseDeviceDTO> purchaseDevices, PaymentMethodTypes paymentMethod)
+        public PurchaseForDetailDTO(int id, string applicationUserName, string applicationUserSurname, string deliveryAddress, DateTime purchaseDate, decimal totalPrice, IList<PurchaseDeviceDTO> purchaseDevices, PaymentMethodTypes paymentMethod)
              : base(applicationUserName, applicationUserSurname, deliveryAddress,paymentMethod, purchaseDate, purchaseDevices )
         {
             Id = id;
